Refactor donors total calculation into modular hooks
